requirements.txt
asgiref==3.8.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
Django==5.1.7
pillow==11.1.0
python-dateutil==2.9.0.post0
six==1.17.0
django-filter==24.2

